+++ CHANGELOG.md
@@ -46,6 +46,15 @@
   introduced in LLVM 21.
 * Add `dilColumn`, `dilIsArtificial`, and `dilCoroSuspendIdx` fields to
   `DILabel'`, which were introduced in LLVM 21.
+* The following debug-related fields have had their types changed from `Word64`
+  to `Maybe (ValMd' lab)`:
+
+  * `DIBasicType'`: `dibtSize`
+  * `DICompositeType'`: `dictSize` and `dictOffset`
+  * `DIDerivedType'`: `didtSize` and `didtOffset`
+
+  This allows them to encode non-constant sizes and offsets (a capability used
+  by Ada, for instance) in LLVM 21 or later.
 * Fix a bug that would cause `indirectbr` statements to be pretty-printed
   incorrectly.